+++ Maddori.Apple/Maddori.Apple/Screen/MyFeedback/MyFeedback/MyFeedbackViewController.swift
@@ -79,29 +79,7 @@ final class MyFeedbackViewController: BaseViewController {
         }
         return collectionView
     }()
-    private lazy var emptyView: UIView = {
-        let view = UIView()
-        view.backgroundColor = .backgroundWhite
-        return view
-    }()
-    private lazy var emptyIcon: UIImageView = {
-        let imageView = UIImageView()
-        imageView.image = ImageLiterals.icPersonCircle
-        imageView.tintColor = .gray700
-        let config = UIImage.SymbolConfiguration(font: .systemFont(ofSize: 50, weight: .thin))
-        imageView.preferredSymbolConfiguration = config
-        return imageView
-    }()
-    private lazy var emptyLabel: UILabel = {
-        let label = UILabel()
-        label.text = TextLiteral.myFeedbackViewControllerEmptyViewLabel
-        label.setLineSpacing(to: 4)
-        label.textAlignment = .center
-        label.numberOfLines = 2
-        label.textColor = .gray700
-        label.font = .body3
-        return label
-    }()
+    private lazy var emptyView = EmptyPersonView()
     
     // MARK: - life cycle
     
@@ -154,19 +132,6 @@ final class MyFeedbackViewController: BaseViewController {
         emptyView.snp.makeConstraints {
             $0.top.equalTo(myFeedbackLabel.snp.bottom)
             $0.leading.trailing.bottom.equalTo(view.safeAreaLayoutGuide)
-        }
-        
-        emptyView.addSubview(emptyIcon)
-        emptyIcon.snp.makeConstraints {
-            $0.centerY.equalToSuperview().offset(-40)
-            $0.centerX.equalToSuperview()
-            $0.width.height.equalTo(60)
-        }
-        
-        emptyView.addSubview(emptyLabel)
-        emptyLabel.snp.makeConstraints {
-            $0.top.equalTo(emptyIcon.snp.bottom).offset(21)
-            $0.centerX.equalToSuperview()
         }
     }
